Move type data to separate shared file
