+++ react-native/react/tracker/bio.render.types.js
@@ -1,0 +1,15 @@
+/* @flow */
+
+export type UserInfo = {
+  fullname: string,
+  followersCount: number,
+  followingCount: number,
+  followsYou: boolean,
+  avatar: ?string,
+  location: string
+}
+
+export type BioProps = {
+  username: ?string,
+  userInfo: ?UserInfo
+}
