+++ react-native/react/tracker/header.render.types.js
@@ -1,0 +1,7 @@
+/* @flow */
+
+export type HeaderProps = {
+  reason: string,
+  onClose: () => void
+}
+
