+++ react-native/react/tracker/proofs.render.types.js
@@ -1,0 +1,18 @@
+/* @flow */
+
+import type {SimpleProofState, SimpleProofMeta} from '../constants/tracker'
+
+export type Proof = {
+  id: string,
+  type: string,
+  state: SimpleProofState,
+  meta: ?SimpleProofMeta,
+  humanUrl: ?string,
+  name: string,
+  color: string
+}
+
+export type ProofsProps = {
+  proofs: Array<Proof>,
+  username: ?string
+}
