+++ react-native/react/tracker/render.types.js
@@ -1,0 +1,13 @@
+/* @flow */
+
+import type {BioProps} from './bio.render.types'
+import type {ActionProps} from './action.render.types'
+import type {HeaderProps} from './header.render.types'
+import type {ProofsProps} from './proofs.render.types'
+
+export type RenderProps = {
+  bioProps: BioProps,
+  actionProps: ActionProps,
+  headerProps: HeaderProps,
+  proofsProps: ProofsProps
+}
